Moved restifyRequestCache method from example to src files
